+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 History
 =======
 
+0.3.0 (May, 2020)
+-------------------------
+* [#10](https://github.com/keyko-io/nevermined-sdk-py/pull/10) Support new authorization methods and use new gateway endpoint for access
+* [#9](https://github.com/keyko-io/nevermined-sdk-py/pull/9) Add docs in github pages
+* [#8](https://github.com/keyko-io/nevermined-sdk-py/pull/8) Upgrade web3py library to last version
+* [#6](https://github.com/keyko-io/nevermined-sdk-py/issues/6) Implementation of the new flow
+
 0.2.0 (April, 2020)
 -------------------------